Implement `PartialEq` for `SingleVersion`
Since it's not clear how to compare our own version struct from maven,
it's marked as `unimplemented`. In order to support this, we need to
create custom version struct for semver version. The `semver` crate
cannot be kept since in fabric mod json file spec, it is specified that
if we can't parse the semver version, we should keep it anyways, but we
  limit the operations that we can apply to it. Clearly, `semver` crate
  doesn't support such functionality