Drop versions support
It has become impossible to maintain the version code, while the custom `semver` version parser was working perfectly, the fabric mod spec tells explicitly that even if it doesn't follow the `semver` format, you should still keep it. Another problem is the `maven` version format, it is so random, that it is simpler to store it as a string.